Refactor `GraphQLService` class and its instantiation.
Previously, with each GraphQLService call, the same HTTP headers had to be
added to each method call, which was the caller's responsibility.
GraphQLService is now made session-based, so that session-specific headers are
added to each GraphQL request on behalf of the caller.

Sending a GraphQL query is now also abstracted separately from the query
construction logic.